requirements.txt
python-dotenv==1.0.1
haystack-ai==2.5.1
haystack-experimental==0.1.0
sentence-transformers==3.0.1 
cohere-haystack==1.1.3
colorama==0.4.6
trafilatura==1.11.0
geopy==2.2.0
pdfminer.six==20201018
pydantic>=1.9.2
jinja2==3.0.2
pdfkit==0.6.1
wkhtmltopdf==0.2
flask==2.0.2
Werkzeug==2.2.2
jsonpickle==2.0.0
pinecone-haystack>=1.2.3